+++ src/country/country_game_data.cpp
@@ -602,7 +602,7 @@ void country_game_data::change_level(const int change)
 			}
 		}
 
-		const int level_experience = defines::get()->get_experience_for_level(this->get_level());
+		const int64_t level_experience = defines::get()->get_experience_for_level(this->get_level());
 		if (this->get_experience() < level_experience) {
 			this->set_experience(level_experience);
 		}
@@ -613,7 +613,7 @@ void country_game_data::change_level(const int change)
 	}
 }
 
-void country_game_data::change_experience(const int change)
+void country_game_data::change_experience(const int64_t change)
 {
 	if (change == 0) {
 		return;
