+++ src/database/defines.cpp
@@ -72,7 +72,7 @@ void defines::process_gsml_scope(const gsml_data &scope)
 	} else if (tag == "experience_per_level") {
 		scope.for_each_property([&](const gsml_property &property) {
 			const int level = std::stoi(property.get_key());
-			const int experience = std::stoi(property.get_value());
+			const int64_t experience = std::stoll(property.get_value());
 
 			this->experience_per_level[level] = experience;
 		});
